requirements.txt
opencv-python
numpy
insightface
faiss-cpu
pyzbar
python-multipart
